+++ src/main/java/ua/edu/sumdu/j2ee/chepiha/eshop/eshop/models/mappers/LocationMapper.java
@@ -10,7 +10,7 @@ public class LocationMapper implements RowMapper<Location> {
 
     public Location mapRow(ResultSet resultSet, int rowNum) throws SQLException {
         Location location = new Location();
-        location.setId(resultSet.getInt("ID"));
+        location.setId(resultSet.getLong("ID"));
         location.setName(resultSet.getString("NAME"));
         location.setAddress(resultSet.getString("ADDRESS"));
         return location;
